requirements.txt
# Core Data Engineering
pyarrow>=14.0.0  # For parquet support
polars>=0.19.0   # High-performance dataframe library
duckdb>=0.9.0    # In-process SQL analytics (Gold layer)

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0  # Used by utils/config.py (BaseSettings)

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
pyyaml>=6.0.0
python-dateutil>=2.8.0

# Logging
loguru>=0.7.0

# SurrealDB (Graph Layer - Task 4)
surrealdb>=0.3.0
pytest-asyncio>=0.21.0
